renamed mock packages aliases
